visual-voice-assistant/requirements.txt
numpy==1.24.2
tqdm==4.65.0
keras==2.11.0
pillow==9.4.0
cli_args_system==1.3
pydot==1.4.2
graphviz==0.20.1
nltk==3.8.1
matplotlib==3.7.1